Impl'd CopyFolderAsync
Added a test.  I think i'll start doing most of the complicated
scenarios (ie, 'copying a folder fails when the destination parent and
name have not changed.') in the Sync versions of the tests. It's more
straightforward, and the underlying logic is the pretty much the same
for both Sync & Async.