Impl'd CopyFile{Async}; Improved error handling
No longer using the 'expected http status code' as a barometer of
success.  Now checking the response.Content for the existence of error
information, and then parsing that out as possible.  I'm still learning
the ins/outs of how error info is returned, so this is a work in
progress.  Any gathered error information is thrown along with the
BoxException.