GetFile (sync) now uses exponential backoff.
A File object needs to have an etag, but this can take some time to
compute on the server.  We now employ an exponential backoff algorithm
to give the server enough -- but not to much -- time to take care of
that.  Need to implement this as well in the GetFileAsync method.
Housekeeping: Removed unused HttpStatusCode parameters from Execute*
signatures.