WHAT THE I DON'T EVEN
Rokay raggy, so this is a very experimental change introducing a UI for
configuring the test credentials and storing them in a local file,
test_info.json, that is .gitignore'd so that people can easily run the
tests and get an auth token stored locally.

Genius or utter stupidity?  Time will tell.