Adding Event/EventEntry model class
Going to have to decide how to deal with the problem that event sources
can be any of files, folders, comments, etc.  Currently just trying to
provide basic, general information (what it is, who created the thing,
etc.)  If the event describes the creation of a comment, then you'll
have to separately fetch that comment to get its message.  We might be
able to do something clever with a custom JSON deserializer, but simple
might be better than clever here...