Start implementing support for shared items
Added BoxManager.GetSharedItem, which will get a shared file or folder
given a shared link url.
Added overload for BoxManager.GetFile that will accept a sharedLinkUrl
in addition to the file id.
RequestAuthenticator will now add the shared link url to the header if
it's provided prior to execution of the request..
BoxRestClient can now set/clear the shared link URL on the
RequestAuthenticator.