Added async folder item limit/offset
This works a little differently than the sync GetFolder/GetItems: if
limit and offset are both left null, then the BoxManager just uses the
default values.  It does not get everything in the folder.  This is
because async operations tend more towards a 'paged' request style, but
I also wanted to introduce some backwards compatibility.